renderer work

+++ src/ray.rs
@@ -15,7 +15,7 @@ impl Ray {
         self.direction
     }
 
-    pub fn at(&self, t:f32) -> Point3 {
+    pub fn at(&self, t:f64) -> Point3 {
         self.origin + t * self.direction
     }
 }
